% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_data}
\alias{load_data}
\title{Load edge list for analysis}
\usage{
load_data(df, names = 1L)
}
\arguments{
\item{df}{edge list}

\item{names}{are we dealing with alphaNumeric (1) or numeric (!1) ids}
}
\description{
This function reads the edge list and creates the network for analysis. It does 
not return anything, it just creates required structures in memory.
}
\examples{
library(igraph)
g <- make_full_graph(5) \%du\% make_full_graph(5) \%du\% make_full_graph(5)
g <- add_edges(g, c(1,6, 1,11, 6, 11))
el = as.data.frame(get.edgelist(g,names=TRUE))
rSpectral::load_data(df=el)
status = rSpectral::spectral(fix_neig=0)
spec   = rSpectral::membership(detach_graph=1)
}
