% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{membership}
\alias{membership}
\title{Return membership data}
\usage{
membership(detach_graph = 1L)
}
\arguments{
\item{detach_graph}{whether you want to keep graph in memory}
}
\value{
membership vector
}
\description{
Return membership data
}
\examples{
library(igraph)
g <- make_full_graph(5) \%du\% make_full_graph(5) \%du\% make_full_graph(5)
g <- add_edges(g, c(1,6, 1,11, 6, 11))
el = as.data.frame(get.edgelist(g,names=TRUE))
rSpectral::load_data(df=el)
status1 = rSpectral::spectral(fix_neig=0)
spec1   = rSpectral::membership(detach_graph=0)
status2 = rSpectral::spectral(fix_neig=1)
spec2   = rSpectral::membership(detach_graph=0)
}
