% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSpectral-package.R
\docType{package}
\encoding{UTF-8}
\name{rSpectral}
\alias{rSpectral}
\alias{rSpectral-package}
\title{rSpectral}
\description{
This package implements Spectral Modularity clustering algorithm for
\code{igraph} and \code{graphNEL} graphs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmclean5/rSpectral}
}

}
\author{
Colin Mclean <Colin.D.Mclean@ed.ac.uk>
}
