% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spectral}
\alias{spectral}
\title{Spectral}
\usage{
spectral(
  Cn_min = 1L,
  tol = 1e-05,
  names = 1L,
  fix_neig = 0L,
  verbose = 0L,
  summary = 0L
)
}
\arguments{
\item{names}{}

\item{verbose}{}

\item{summary}{}

\item{DF}{}

\item{CnMIN}{}

\item{TOL}{}

\item{fixNeig}{}
}
\description{
Spectral
}
