% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSpectral-package.R
\docType{package}
\encoding{UTF-8}
\name{rSpectral}
\alias{rSpectral}
\alias{rSpectral-package}
\title{rSpectral}
\description{
This package implements the Spectral Modularity clustering algorithm for
\code{\link{igraph}} and \code{\link[graph]{graphNEL}} graphs. The algorithm 
was proposed in \insertCite{PhysRevE.74.036104}{rSpectral} and an example
of its application to the real biological network could be found in
\insertCite{Roy:2018Regional}{rSpectral}.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmclean5/rSpectral}
  \item Report bugs at \url{https://github.com/cmclean5/rSpectral/issues/}
}

}
\author{
Colin Mclean <Colin.D.Mclean@ed.ac.uk>
}
