% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spectral}
\alias{spectral}
\title{Spectral modularity clustering}
\usage{
spectral(Cn_min = 1L, tol = 1e-05, names = 1L, fix_neig = 0L)
}
\arguments{
\item{Cn_min}{minimum cluster size}

\item{tol}{tolerance}

\item{names}{are we dealing with alphaNumeric (1) or numeric (!1) ids}

\item{fix_neig}{wether fixing neighbouring nodes found in same community}
}
\value{
status (does it really return something?)
}
\description{
This function does the actual job of clustering. Two major parameter are
\code{Cn_min}, which controls the minimal size of the cluster, 
and \code{fix_neig}, which can reduce execution time by factor 2 to 5 by
checking only boundary nodes of already found clusters. Optimal set of 
parameters could be obtained by trying different combinations of values and
using \code{detach_graph=0} when getting communities by calling 
\code{\link{membership}}.
}
\examples{
library(igraph)
g <- make_full_graph(5) \%du\% make_full_graph(5) \%du\% make_full_graph(5)
g <- add_edges(g, c(1,6, 1,11, 6, 11))
el = as.data.frame(get.edgelist(g,names=TRUE))
rSpectral::load_data(df=el)
status = rSpectral::spectral(fix_neig=0)
spec   = rSpectral::membership(detach_graph=1)
V(g)$color[as.numeric(spec$ID)]<-RColorBrewer::brewer.pal(max(spec$K),'Set1')[spec$K]
plot(g)
}
